--- a/docs/source/changelog.rst
+++ b/docs/source/changelog.rst
@@ -3,6 +3,11 @@
 ChangeLog
 #########
 
+Version 0.142.1
+===============
+
+* Speed up `poseTransformPoints` by taking advantage of contiguous numpy array when extracting values from inputs as well as caching intermediate values for transform computation.
+
 Version 0.142.0
 ===============